requirements.txt
# Telegram
aiogram==3.13.1
telethon==1.36.0

# OCR
rapidocr-onnxruntime==1.3.25
pillow==10.4.0
numpy==1.26.4

# Utils
python-dotenv==1.0.1
pytz==2024.1

# Tests
pytest==8.3.3
